+++ NEWS.md
@@ -5,6 +5,8 @@
 - New biocpkgversion and githubpkgversion functions <2017-08-25 Fri>
 - biocVersions function for Bioconducor release and devel versions
   (moved over from cputols) <2017-09-07 Thu>
+- Add a type (software, experiment, annotation) to biocpkgversion
+  <2017-09-08 Fri>
 
 ## biocpkgs 0.1.0